# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication, Authorization, and Accounting Service with PostgreSQL-based RBAC",
        "title": "AAA Service API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v2/auth/login": {
            "post": {
                "description": "Authenticate user with username and password and MFA",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "User login (V2)",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.LoginSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Logout user and invalidate tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "User logout (V2)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.LogoutSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/refresh": {
            "post": {
                "description": "Refresh access token using refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Refresh access token (V2)",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.RefreshTokenSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    }
                }
            }
        },
        "/api/v2/auth/register": {
            "post": {
                "description": "Register a new user account with enhanced validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "User registration (V2)",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.RegisterSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponseSwagger"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "mpin",
                "refresh_token"
            ],
            "properties": {
                "mpin": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "responses.AddressInfo": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "full_address": {
                    "type": "string"
                },
                "house": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "landmark": {
                    "type": "string"
                },
                "pincode": {
                    "type": "string"
                },
                "post_office": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "subdistrict": {
                    "type": "string"
                },
                "vtc": {
                    "description": "Village/Town/City",
                    "type": "string"
                }
            }
        },
...
        "responses.ContactInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_primary": {
                    "type": "boolean"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "responses.ErrorResponseSwagger": {
            "description": "Standard error response structure",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "error": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid input data"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456789"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "responses.LoginResponseData": {
            "description": "Login response data containing tokens and user info",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 86400
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                },
                "user": {
                    "$ref": "#/definitions/responses.UserInfo"
                }
            }
        },
        "responses.LoginSuccessResponse": {
            "description": "Successful login response with tokens and user info",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.LoginResponseData"
                },
                "message": {
                    "type": "string",
                    "example": "Login successful"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456789"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "responses.LogoutSuccessResponse": {
            "description": "Successful logout response",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Logout successful"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456789"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "responses.RefreshTokenResponseData": {
            "description": "Refresh token response data",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 86400
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
        "responses.RefreshTokenSuccessResponse": {
            "description": "Successful token refresh response",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.RefreshTokenResponseData"
                },
                "message": {
                    "type": "string",
                    "example": "Token refreshed successfully"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456789"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "responses.RegisterResponseData": {
            "description": "Registration response data containing user info",
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/responses.UserInfoSwagger"
                }
            }
        },
        "responses.RegisterSuccessResponse": {
            "description": "Successful registration response with user info",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.RegisterResponseData"
                },
                "message": {
                    "type": "string",
                    "example": "User registered successfully"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456789"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "responses.RoleDetail": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "responses.UserInfo": {
            "type": "object",
            "properties": {
                "contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ContactInfo"
                    }
                },
                "country_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "has_mpin": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_validated": {
                    "type": "boolean"
                },
                "phone_number": {
                    "type": "string"
                },
                "profile": {
                    "$ref": "#/definitions/responses.UserProfileInfo"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.UserRoleDetail"
                    }
                },
                "status": {
                    "type": "string"
                },
                "tokens": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "responses.UserInfoSwagger": {
            "description": "Enhanced user information structure with roles and profile",
            "type": "object",
            "properties": {
                "contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ContactInfo"
                    }
                },
                "country_code": {
                    "type": "string",
                    "example": "US"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "has_mpin": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "string",
                    "example": "USER123456789"
                },
                "is_validated": {
                    "type": "boolean",
                    "example": true
                },
                "phone_number": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "profile": {
                    "$ref": "#/definitions/responses.UserProfileInfo"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.UserRoleDetailInfo"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "tokens": {
                    "type": "integer",
                    "example": 100
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "responses.UserProfileInfo": {
            "type": "object",
            "properties": {
                "aadhaar_number": {
                    "type": "string"
                },
                "address": {
                    "$ref": "#/definitions/responses.AddressInfo"
                },
                "care_of": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email_hash": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "share_code": {
                    "type": "string"
                },
                "year_of_birth": {
                    "type": "string"
                }
            }
        },
        "responses.UserRoleDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "role": {
                    "$ref": "#/definitions/responses.RoleDetail"
                },
                "role_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "responses.UserRoleDetailInfo": {
            "description": "User role detail information structure",
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "USERROLE123456789"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "role": {
                    "$ref": "#/definitions/responses.RoleDetail"
                },
                "role_id": {
                    "type": "string",
                    "example": "ROLE123456789"
                },
                "user_id": {
                    "type": "string",
                    "example": "USER123456789"
                }
            }
        },
...
        "services.LoginRequest": {
            "type": "object",
            "required": [
                "country_code",
                "password",
                "phone_number"
            ],
            "properties": {
                "country_code": {
                    "type": "string"
                },
                "mfa_code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "services.RegisterRequest": {
            "type": "object",
            "required": [
                "country_code",
                "password",
                "phone_number"
            ],
            "properties": {
                "country_code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "phone_number": {
                    "type": "string"
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}